Separate variable discovery to new op manager (Discoverer)
